+++ src/mfx/dsp/ctrl/lfo/OscParabola.cpp
@@ -116,7 +116,7 @@ bool	OscParabola::do_is_using_variation (int /*param*/) const
 
 
 
-void	OscParabola::do_tick (long nbr_spl)
+void	OscParabola::do_tick (int nbr_spl)
 {
 	_phase_gen.tick (nbr_spl);
 }
