+++ src/mfx/dsp/ctrl/lfo/OscSquare.cpp
@@ -119,7 +119,7 @@ bool	OscSquare::do_is_using_variation (int /*param*/) const
 
 
 
-void	OscSquare::do_tick (long nbr_spl)
+void	OscSquare::do_tick (int nbr_spl)
 {
 	_phase_gen.tick (nbr_spl);
 }
