+++ src/mfx/dsp/ctrl/lfo/OscVariSlope.cpp
@@ -135,7 +135,7 @@ bool	OscVariSlope::do_is_using_variation (int param) const
 
 
 
-void	OscVariSlope::do_tick (long nbr_spl)
+void	OscVariSlope::do_tick (int nbr_spl)
 {
 	_phase_gen.tick (nbr_spl);
 }
