+++ src/mfx/dsp/dyn/EnvFollowerRmsSimple.cpp
@@ -78,18 +78,18 @@ void	EnvFollowerRmsSimple::set_time (float t)
 
 
 
-void	EnvFollowerRmsSimple::process_block (float out_ptr [], const float in_ptr [], long nbr_spl)
+void	EnvFollowerRmsSimple::process_block (float out_ptr [], const float in_ptr [], int nbr_spl)
 {
 	assert (out_ptr != 0);
 	assert (in_ptr != 0);
 	assert (nbr_spl > 0);
 
-	float				state = _state;
+	float          state = _state;
 
-	long				pos = 0;
+	int            pos = 0;
 	do
 	{
-		const float		x = in_ptr [pos];
+		const float    x = in_ptr [pos];
 		process_sample_internal (state, x);
 		out_ptr [pos] = float (sqrt (state));
 
@@ -102,18 +102,18 @@ void	EnvFollowerRmsSimple::process_block (float out_ptr [], const float in_ptr [
 
 
 
-void	EnvFollowerRmsSimple::process_block_no_sqrt (float out_ptr [], const float in_ptr [], long nbr_spl)
+void	EnvFollowerRmsSimple::process_block_no_sqrt (float out_ptr [], const float in_ptr [], int nbr_spl)
 {
 	assert (out_ptr != 0);
 	assert (in_ptr != 0);
 	assert (nbr_spl > 0);
 
-	float				state = _state;
+	float          state = _state;
 
-	long				pos = 0;
+	int            pos = 0;
 	do
 	{
-		const float		x = in_ptr [pos];
+		const float    x = in_ptr [pos];
 		process_sample_internal (state, x);
 		out_ptr [pos] = state;
 
@@ -128,18 +128,18 @@ void	EnvFollowerRmsSimple::process_block_no_sqrt (float out_ptr [], const float 
 
 // Input is not squared
 // Output is not square-rooted
-void	EnvFollowerRmsSimple::process_block_raw (float out_ptr [], const float in_ptr [], long nbr_spl)
+void	EnvFollowerRmsSimple::process_block_raw (float out_ptr [], const float in_ptr [], int nbr_spl)
 {
 	assert (out_ptr != 0);
 	assert (in_ptr != 0);
 	assert (nbr_spl > 0);
 
-	float				state = _state;
+	float          state = _state;
 
-	long				pos = 0;
+	int            pos = 0;
 	do
 	{
-		const float		x2 = in_ptr [pos];
+		const float    x2 = in_ptr [pos];
 		assert (x2 >= 0);
 		process_sample_internal_no_sq (state, x2);
 		out_ptr [pos] = state;
@@ -153,17 +153,17 @@ void	EnvFollowerRmsSimple::process_block_raw (float out_ptr [], const float in_p
 
 
 
-float	EnvFollowerRmsSimple::analyse_block (const float data_ptr [], long nbr_spl)
+float	EnvFollowerRmsSimple::analyse_block (const float data_ptr [], int nbr_spl)
 {
 	assert (data_ptr != 0);
 	assert (nbr_spl > 0);
 
-	float				state = _state;
+	float          state = _state;
 
-	long				pos = 0;
+	int            pos = 0;
 	do
 	{
-		const float		x = data_ptr [pos];
+		const float    x = data_ptr [pos];
 		process_sample_internal (state, x);
 
 		++ pos;
@@ -177,14 +177,14 @@ float	EnvFollowerRmsSimple::analyse_block (const float data_ptr [], long nbr_spl
 
 
 
-float	EnvFollowerRmsSimple::analyse_block_no_sqrt (const float data_ptr [], long nbr_spl)
+float	EnvFollowerRmsSimple::analyse_block_no_sqrt (const float data_ptr [], int nbr_spl)
 {
 	assert (data_ptr != 0);
 	assert (nbr_spl > 0);
 
-	float				state = _state;
+	float          state = _state;
 
-	long				pos = 0;
+	int            pos = 0;
 	do
 	{
 		const float		x = data_ptr [pos];
@@ -203,17 +203,17 @@ float	EnvFollowerRmsSimple::analyse_block_no_sqrt (const float data_ptr [], long
 
 // Input is not squared
 // Output is not square-rooted
-float	EnvFollowerRmsSimple::analyse_block_raw (const float data_ptr [], long nbr_spl)
+float	EnvFollowerRmsSimple::analyse_block_raw (const float data_ptr [], int nbr_spl)
 {
 	assert (data_ptr != 0);
 	assert (nbr_spl > 0);
 
-	float				state = _state;
+	float          state = _state;
 
-	long				pos = 0;
+	int            pos = 0;
 	do
 	{
-		const float		x2 = data_ptr [pos];
+		const float    x2 = data_ptr [pos];
 		assert (x2 >= 0);
 		process_sample_internal_no_sq (state, x2);
 
@@ -230,12 +230,12 @@ float	EnvFollowerRmsSimple::analyse_block_raw (const float data_ptr [], long nbr
 
 // Input is not squared
 // Output is not square-rooted
-float	EnvFollowerRmsSimple::analyse_block_raw_cst (float x2, long nbr_spl)
+float	EnvFollowerRmsSimple::analyse_block_raw_cst (float x2, int nbr_spl)
 {
 	assert (x2 >= 0);
 	assert (nbr_spl > 0);
 
-	const float		delta = x2 - _state;
+	const float    delta = x2 - _state;
 
 	float          coef_block;
 	if (nbr_spl < 100000)
